requirements.txt
boto3==1.1.2
botocore==1.2.0
colorama==0.3.3
docutils==0.12
Flask==0.10.1
futures==2.2.0
itsdangerous==0.24
Jinja2==2.8
jmespath==0.7.1
MarkupSafe==0.23
numpy==1.9.2
Pillow==2.9.0
pyasn1==0.1.8
pyparsing==2.0.3
python-dateutil==2.4.2
pytz==2015.4
requests==2.7.0
rsa==3.1.4
six==1.9.0
Werkzeug==0.10.4
wheel==0.24.0
